requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Kafka
kafka-python==2.0.2

# Spark
pyspark==3.4.1

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# ML & Data Processing
joblib>=1.3.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0